+++ src/models/folders.ts
@@ -103,6 +103,11 @@ export interface ListFolderQueryParams extends ListQueryParams {
   parentId?: string;
 
   /**
+   * (Microsoft only) When true, Nylas includes hidden folders in its response.
+   */
+  includeHiddenFolders?: boolean;
+
+  /**
    * (Microsoft only) If true, retrieves folders from a single-level hierarchy only.
    * If false, retrieves folders across a multi-level hierarchy.
    * @default false